ftp://ftp.linux.ncsu.edu/mirror/ftp.redhat.com/pub/redhat/linux/enterprise/4/en/os/i386/SRPMS/xscreensaver-4.18-5.rhel4.2.src.rpm -rw-r--r-- 1 zblaxell zblaxell 4271312 Feb 27 12:21 xscreensaver-4.18-5.rhel4.2.src.rpm 9f48ac2504ac26a64982fb46c395b7634014b727  xscreensaver-4.18-5.rhel4.2.src.rpm
xscreensaver-4.18.tar.gz
-rw-rw-r-- 2 zblaxell zblaxell 4252004 Aug 31  2004 xscreensaver-4.18.tar.gz
8f7fdf53dbcc3cc96c5d9d4165c66289859d80c5  xscreensaver-4.18.tar.gz

+++ hacks/webcollage-helper.c
@@ -384,7 +384,7 @@ write_pixbuf (GdkPixbuf *pb, const char *file)
           perror (buf);
           exit (1);
         }
-      fprintf (stderr, " %luK\n", (st.st_size + 1023) / 1024);
+      fprintf (stderr, " %luK\n", ((unsigned long) st.st_size + 1023) / 1024);
     }
 
   fclose (out);
